--- app/helpers/r_schedules_helper.rb
@@ -1,2 +1,0 @@
-module RSchedulesHelper
-end
